Add tests including public headers
These tests check whether the inclusion of segyio headers that are part
of segyio's public interface prevent other projects from compiling. We
had cases in which our public facing C and C++ headers did not define
all required include. This could lead to failures depending on the
compiler and what other includes were and in which order when a segyio
header was included..